SQS Plugin
Combine case tests - move queue test into api test

https://eaflood.atlassian.net/browse/FUT-390
